refactor(report): move output related components to dedicated folder
