+++ src/Reportify/Commands/ReportCommand.cs
@@ -2,6 +2,7 @@ using System.Reflection;
 using Reportify.Configuration.Validation;
 using Reportify.Extensions;
 using Reportify.Report;
+using Reportify.Report.Output;
 using Spectre.Console;
 using Spectre.Console.Cli;
